Fix sort and add extreme tests.
The sort was incorrect due to incorrect ranges for alphabets,
which was missing 'z'.

The new tests are quite comprehensive, and test for sort equivalence
against the libapt implementation, by using a sorted list of versions.

We shuffle this list, sort it again, and then check it line by line
against our known sorted version. This might not match for a few lines
at a time, where the versions are equivalent, and the sort order is
undeterministic. In such cases, we validate these exceptions (1500 or so)
as being equivalent in our comparison as well.